{colombo} norming emissions to mg/s and using double in division
git-svn-id: http://svn.code.sf.net/p/sumo/code/branches/colombo@14481 29f149b2-0c63-11e1-8f5c-a347f6a9acd0

+++ src/utils/emissions/PHEMConstants.h
@@ -23,6 +23,6 @@
 
 const double GRAVITY_CONST = 9.81;
 const double AIR_DENSITY_CONST = 1.182;
-const int SECONDS_PER_HOUR = 3600;
+const double SECONDS_PER_HOUR = 3600;
 
 #endif
